+++ tests/testthat.R
@@ -1,14 +1,15 @@
 library(testthat)
 library(stockfish)
 
+# Detect Solaris
 is_solaris <- function() {
   grepl("SunOS", Sys.info()["sysname"])
 }
 
-is_m1 <- function() {
-  !R.version$arch == "x86_64"
-}
-
-if (!is_solaris() && !is_m1()) {
+# Solaris users will have to install 'Stockfish 11' manually depending on their
+# compiler (e.g. the bundled version was working on R-hub, but not on CRAN).
+# Since the tests are supposed to guarantee that the bundled version installed
+# correctly, they are skipped.
+if (!is_solaris()) {
   test_check("stockfish")
 }